# TSDB: optimize series creation on PreCreation() failure