In Cortex we use `SeriesCallback` and we have edge cases where the number of series failing the `PreCreation()` is very high. When this happens, we see an higher CPU utilization and memory allocation caused by `*memSeries` being created before checking if the `PreCreation()` callback passes.

This PR proposed a fix for that.

## Benchmark

```
name                                          old time/op    new time/op    delta
HeadStripeSeriesCreate-12                       1.79µs ±17%    1.84µs ±16%     ~     (p=0.421 n=5+5)
HeadStripeSeriesCreateParallel-12               1.60µs ± 2%    1.76µs ±11%     ~     (p=0.286 n=4+5)
HeadStripeSeriesCreate_PreCreationFailure-12    1.48µs ± 2%    1.27µs ± 0%  -14.50%  (p=0.008 n=5+5)

name                                          old alloc/op   new alloc/op   delta
HeadStripeSeriesCreate-12                         783B ± 2%      813B ±13%     ~     (p=0.246 n=5+5)
HeadStripeSeriesCreateParallel-12                 841B ±23%      890B ±16%     ~     (p=0.730 n=5+5)
HeadStripeSeriesCreate_PreCreationFailure-12      699B ± 0%      395B ± 0%  -43.49%  (p=0.000 n=5+4)

name                                          old allocs/op  new allocs/op  delta
HeadStripeSeriesCreate-12                         8.00 ± 0%      8.00 ± 0%     ~     (all equal)
HeadStripeSeriesCreateParallel-12                 8.00 ± 0%      8.00 ± 0%     ~     (all equal)
HeadStripeSeriesCreate_PreCreationFailure-12      9.00 ± 0%      6.00 ± 0%  -33.33%  (p=0.008 n=5+5)
```